3.4 Update
Now we've added the feature of finding a specific word in the .java file. Press Ctrl+F if you have opened a .java file , and the find dialog will be shown , allowing you find words in the current Java file.

We've also added the feature of saving either the whole Jar file or the Java file you are viewing , you can press Ctrl+S to save the whole Jar file , and press Ctrl+Shift+S to save the current Java file.
For the feature above , we've added the feature of saving the whole jar file automatically.You can enable the feature through the menu item that is in the submenu of "Settings".If the feature is enabled , you can find the decompiled jar file in the folder "\decompile" in the folder which the source jar file is in.
Besides , you can decompile the Jar file through the MRU(Most Recently Used) file list now.This time the application will decompile quickly thanks to the cache it saved.

Some small adjustments aren't metioned.

Fixed:
Now the application will check the existance of Java Runtime Environment asynchronously.
Improved the speed of the File View.

P.S.: I forgot to save the source code of version 3.4,so the source code published belongs to test version of 3.4.1.